Add initial 'language' definition for irc logs, extend color scheme.
The 'pack' version is generated from a highlightjs tree,
and must be used (as opposed to minimized source, this is packed)
in order to have fields/etc align with current highlightjs.

Will likely replace our highlightjs entirely in the near future,
incorporating this highlighting, but will keep the unpacked
language definition for reference/updates.